+++ server/datastore/mysql/migrations/tables/20240212111533_AddPINToHostMDMActions.go
@@ -10,6 +10,7 @@ func init() {
 }
 
 func Up_20240212111533(tx *sql.Tx) error {
+	// Idempotent migration.
 	// going with a VARCHAR instead of a number because the leading zeros are
 	// important in a PIN. Being a VARCHAR will also make it easy to make larger
 	// if needed in the future.
@@ -17,13 +18,15 @@ func Up_20240212111533(tx *sql.Tx) error {
 	// An unlock_ref field is also necessary for Windows/Linux where unlocking is
 	// done via a script, so we need a reference to that script's execution uuid
 	// as we already have for lock_ref and wipe_ref.
-	stmt := `ALTER TABLE host_mdm_actions
+	if !columnsExists(tx, "host_mdm_actions", "unlock_pin", "unlock_ref") {
+		stmt := `ALTER TABLE host_mdm_actions
 		ADD COLUMN unlock_pin VARCHAR(6) NULL,
 		ADD COLUMN unlock_ref VARCHAR(36) NULL,
 		DROP COLUMN suspended
 `
-	if _, err := tx.Exec(stmt); err != nil {
-		return fmt.Errorf("alter table host_mdm_actions: %w", err)
+		if _, err := tx.Exec(stmt); err != nil {
+			return fmt.Errorf("alter table host_mdm_actions: %w", err)
+		}
 	}
 	return nil
 }
